2025-12-02_taller_4/python/requirements.txt
# Detection and Segmentation Subsystem Dependencies
# Python 3.8+

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Object Detection - YOLO
ultralytics>=8.0.0

# Segmentation - SAM
segment-anything
git+https://github.com/facebookresearch/segment-anything.git

# Image Processing
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Backend API
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Video processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Metrics and monitoring
psutil>=5.9.0
gpustat>=1.1.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
